+++ src/main/java/vv/intertrain/vidmot/UndirbuaController.java
@@ -5,12 +5,144 @@ import javafx.scene.control.TextField;
 import vv.intertrain.vinnsla.ChatMode;
 import vv.intertrain.vinnsla.InterviewBot;
 
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+import javafx.scene.control.ListView;
+import javafx.scene.control.TextArea;
+// import vinnsla.Spurningar;
+
+import static vv.intertrain.vidmot.InterviewApplication.interview;
+
+import java.util.Map;
+import java.util.Optional;
+
+public class UndirbuaController {
+    @FXML
+    private Label fxFjoldiSvara;
+    @FXML
+    private ListView<String> fxFlokkar;
+    @FXML
+    private ListView<String> fxSpurningar;
+    @FXML
+    private TextArea fxSpurningaLog;
+
+    private String valinnFlokkur;
+    private String valinSpurning;
+    // private Spurningar spurningar;
+    private final ObservableList<String> spurningaLog = FXCollections.observableArrayList();
+    private final StringProperty spurningaLogProperty = new SimpleStringProperty("");
+
+    @FXML
+    protected void onSvara() {
+        // sækja valið item úr spurningalistanum, bara virkja takkann ef eitthvað er valið
+        updateValinSpurning();
+        SvarDialogController svarDialogController = new SvarDialogController(valinSpurning);
+        Optional<String> utkoma = svarDialogController.showAndWait();
+
+        String svar = svarDialogController.getSvar();
+
+        if (svar != null && !svar.isEmpty()) { // ef notandi svaraði
+            // hækkar fjölda svara um 1
+            //int fjoldi = Integer.parseInt(spurningar.getFjoldiSvaradraSpurninga());
+            //spurningar.setFjoldiSvaradraSpurninga(String.valueOf(++fjoldi));
+            // bætir spurningu í log
+            spurningaLog.add(valinSpurning);
+            updateSpurningaLog();
+        }
+    }
+
+    @FXML
+    protected void onTilBaka() {
+        // fara til baka í Velkominn
+        ViewSwitcher.switchTo(View.VELKOMINN, true);
+    }
+
+    @FXML
+    protected void onHaetta() {
+        // skipta yfir í kveðju
+        ViewSwitcher.switchTo(View.KVEDJA, true);
+    }
+
+    protected String getValinnFlokkur() {
+        return this.valinnFlokkur;
+    }
+
+    protected void updateValinnFlokkur() {
+        this.valinnFlokkur = fxFlokkar.getSelectionModel().getSelectedItem();
+    }
+
+    protected String getValinSpurning() {
+        return this.valinSpurning;
+    }
+
+    protected void updateValinSpurning() {
+        this.valinSpurning = fxSpurningar.getSelectionModel().getSelectedItem();
+    }
+
+    private void updateSpurningaLog() {
+        String logText = String.join("\n", spurningaLog);
+        spurningaLogProperty.set(logText);
+    }
+
+    public void initialize() {
+
+        String nafn = ChatboxController.getNafn();
+        String starf = ChatboxController.getStarf();
+        String fyrirtaeki = ChatboxController.getFyrirtaeki();
+
+        interview = new InterviewBot(nafn, fyrirtaeki, starf, 10, ChatMode.PREPARATION);
+
+
+        // this.spurningar = new Spurningar();
+
+        // bindur fjölda svara við viðmótshlut
+        // fxFjoldiSvara.textProperty().bind(spurningar.getFjoldiSvaradraSpurningaProperty());
+        // bindur spurningalog við viðmótshlut
+        fxSpurningaLog.textProperty().bind(spurningaLogProperty);
+
+        //fxFlokkar.setItems(spurningar.getFlokkar());
+
+        // if (!spurningar.getFlokkar().isEmpty()) {
+            updateValinnFlokkur();
+        }
+
+        /*
+        fxFlokkar.getSelectionModel().selectedIndexProperty().addListener((obs, old, newIndex) -> {
+            if (newIndex != null && newIndex.intValue() > -1) {
+                updateValinnFlokkur();
+
+                Map<String, ObservableList<String>> spurningaListi = spurningar.getSpurningaListi(this.getValinnFlokkur());
+                fxSpurningar.setItems(spurningaListi.get(this.getValinnFlokkur()));
+                fxSpurningar.getSelectionModel().selectFirst();
+                updateValinSpurning();
+            }
+        });
+        */
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+/*
 public class UndirbuaController {
     @FXML private ChatboxController chatBoxController;
     @FXML private TextField inntak;
-
-    // TEMP FIX COMMENT
-    private InterviewBot prep;
 
     @FXML
     public void initialize() throws Exception {
@@ -60,3 +192,4 @@ public class UndirbuaController {
         ViewSwitcher.switchTo(View.KVEDJA, true);
     }
 }
+ */
