+++ src/main/java/vv/intertrain/vidmot/View.java
@@ -7,7 +7,8 @@ package vv.intertrain.vidmot;
 public enum View {
     VELKOMINN("velkominn-view.fxml"),
     VIDTAL("vidtal-view.fxml"),
-    UNDIRBUA("undirbua-view.fxml"),
+    // UNDIRBUA("undirbua-view.fxml"),
+    UNDIRBUA("spurningar-view.fxml"),
     KVEDJA("kvedja-view.fxml");
 
     private final String fileName;
