+++ src/main/java/vv/intertrain/vinnsla/InterviewPreparer.java
@@ -1,0 +1,4 @@
+package vv.intertrain.vinnsla;
+
+public class InterviewPreparer {
+}
